worker_core1: run squared-FFT CFO before RRC (gr-iridium order)
Mirror burst_downmix_impl.cc::process_next_frame (lines 506-594):
  D13 start → squared-FFT CFO → freq-correct → RRC → UW correlator.

Previously the worker applied RRC first, then ran uw_correlator_find
whose internal cfo_fine_estimate ran post-RRC. The 2f_c tone of a
band-edge carrier is outside the RRC passband; the squared-FFT picked
DC leakage instead.

Also fixes an int32-overflow regression introduced in the previous
commit's adaptive squared-FFT scaling: (sq_re * w) where sq_re reaches
2^22 and w is Q15 (2^15) can hit 2^37, overflowing int32. Cast w to
int64 before the multiply.

All host CFO tests pass:
  uw_correlator_cfo (synthetic, 10 cases including BPSK + noise)
  uw_correlator_albq (real-RF DL fixture)
  uw_correlator_albq_ul (real-RF UL fixture)
  demod_albq (DL end-to-end, 0% BER)

Firmware build succeeds (5.2 MB binary, 57% partition free).

+++ common/iridium_decoder/uw_correlator.c
@@ -599,14 +599,16 @@ static float cfo_fine_estimate(const int16_t *burst_2sps, int n_complex,
     int sq_shift = 0;
     while ((peak_abs >> sq_shift) > (1 << 11)) sq_shift++;
     for (int i = 0; i < n_in; i++) {
-        int32_t w = (int32_t)win[i];
+        int64_t w = (int64_t)win[i];
         int32_t r = (int32_t)burst_2sps[(start + i) * 2 + 0] >> sq_shift;
         int32_t m = (int32_t)burst_2sps[(start + i) * 2 + 1] >> sq_shift;
         int32_t sq_re = r * r - m * m;
         int32_t sq_im = 2 * r * m;
+        // sq_re/sq_im can reach 2^22, win is Q15 (≤ 2^15), so the
+        // product can hit 2^37 — must use int64 to avoid overflow.
         // No int16 cast — FFT input is int32 BFP, full int32 range OK.
-        re[i] = (sq_re * w) >> 15;
-        im[i] = (sq_im * w) >> 15;
+        re[i] = (int32_t)((sq_re * w) >> 15);
+        im[i] = (int32_t)((sq_im * w) >> 15);
     }
 
     (void)cfo_fft_q15(re, im);
